requirements.txt
fastapi==0.111.0
uvicorn==0.23.1
google-auth==2.23.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.85.0
requests==2.32.3
Jinja2==3.1.6
python-multipart==0.0.9
